Disassembler - Added an offset to account for the beginning memory
Offset is of value 13, plus the initial starting memory of whatever Disassembler is given.
Only changed E1.txt, need to change the other 2.
Cleaned up the printing by adding new line.
Overall, added the starting and ending naming, with main and the stack.
However, need to account for potential arrays if being used as well as if given the entire object code file with main to stack.
Stack offset is 300.