Quick Changes and testing
Issue Found:
the mov's seem to be incorrectly set up, need more testing to ensure it is correct.
Specifically irmovq is not working correctly!
Also began setting up the checkStart function which checks if the object code contains the start of the file, as in calling in main with a halt and setting up the stack pointer.